+++ samples/sampler/stories/default/Input.js
@@ -1,6 +1,6 @@
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import {action} from '@enact/storybook-utils/addons/actions';
-import {boolean, select, text} from '@enact/storybook-utils/addons/knobs';
+import {boolean, select, text} from '@enact/storybook-utils/addons/controls';
 import {Input, InputBase} from '@enact/agate/Input';
 
 import {iconList} from './util/icons';
@@ -18,26 +18,39 @@ export default {
 	component: 'Input'
 };
 
-export const _Input = () => (
+export const _Input = (args) => (
 	<Input
-		autoFocus={boolean('autoFocus', Config)}
-		clearButton={boolean('clearButton', Config)}
-		clearIcon={select('clearIcon', iconNames, Config)}
-		disabled={boolean('disabled', Config)}
-		dismissOnEnter={boolean('dismissOnEnter', Config)}
-		iconAfter={select('iconAfter', iconNames, Config)}
-		iconBefore={select('iconBefore', iconNames, Config)}
-		invalid={boolean('invalid', Config)}
-		invalidMessage={text('invalidMessage', Config)}
+		autoFocus={args['autoFocus']}
+		clearButton={args['clearButton']}
+		clearIcon={args['clearIcon']}
+		disabled={args['disabled']}
+		dismissOnEnter={args['dismissOnEnter']}
+		iconAfter={args['iconAfter']}
+		iconBefore={args['iconBefore']}
+		invalid={args['invalid']}
+		invalidMessage={args['invalidMessage']}
 		onBeforeChange={action('onBeforeChange')}
 		onBlur={action('onBlur')}
 		onChange={action('onChange')}
 		onFocus={action('onFocus')}
-		placeholder={text('placeholder', Config, 'Input text here')}
-		size={select('size', ['small', 'large'], Config, 'large')}
-		type={select('type', prop.type, Config)}
+		placeholder={args['placeholder']}
+		size={args['size']}
+		type={args['type']}
 	/>
 );
+
+boolean('autoFocus', _Input, Config);
+boolean('clearButton', _Input, Config);
+select('clearIcon', _Input, iconNames, Config);
+boolean('disabled', _Input, Config);
+boolean('dismissOnEnter', _Input, Config);
+select('iconAfter', _Input, iconNames, Config);
+select('iconBefore', _Input, iconNames, Config);
+boolean('invalid', _Input, Config);
+text('invalidMessage', _Input, Config);
+text('placeholder', _Input, Config, 'Input text here');
+select('size', _Input, ['small', 'large'], Config, 'large');
+select('type', _Input, prop.type, Config);
 
 _Input.storyName = 'Input';
 _Input.parameters = {