+++ samples/sampler/stories/qa/Dropdown.js
@@ -1,12 +1,12 @@
-import Group from '@enact/ui/Group';
-import {mergeComponentMetadata} from '@enact/storybook-utils';
-import {boolean, number, select, text} from  '@enact/storybook-utils/addons/knobs';
-import {action} from '@enact/storybook-utils/addons/actions';
-import {Component} from 'react';
 import Button from '@enact/agate/Button';
 import Dropdown, {DropdownBase} from '@enact/agate/Dropdown';
 import Heading from '@enact/agate/Heading';
 import Scroller from '@enact/agate/Scroller';
+import {mergeComponentMetadata} from '@enact/storybook-utils';
+import {action} from '@enact/storybook-utils/addons/actions';
+import {boolean, range, select, text} from '@enact/storybook-utils/addons/controls';
+import Group from '@enact/ui/Group';
+import {useCallback, useState} from 'react';
 
 Dropdown.displayName = 'Dropdown';
 const Config = mergeComponentMetadata('Dropdown', Dropdown, DropdownBase);
@@ -18,166 +18,173 @@ const list = [
 	{children: 'hello 3', 'key': 'key3', 'aria-label': 'aria 3'}
 ];
 
-class AutoDismissDropdown extends Component {
-	constructor (props) {
-		super(props);
-		this.state = {
-			open: true
-		};
-	}
+const AutoDismissDropdown = () => {
+	const [open, setOpen] = useState(true);
 
-	handleClose = () => {
-		this.setState({open: false});
-	};
+	const handleClose = useCallback(() => {
+		setOpen(false);
+	}, []);
 
-	handleOpen = () => {
-		this.setState({open: true});
-	};
+	const handleOpen = useCallback(() => {
+		setOpen(true);
+	}, []);
 
-	render () {
-		return (
-			<div>
-				<Heading>Click in the blank area of the viewport to dismiss the Dropdown</Heading>
-				<Dropdown
-					onClose={this.handleClose}
-					onOpen={this.handleOpen}
-					open={this.state.open} // initial value is true
-				>
-					{['test1', 'test2', 'test3']}
-				</Dropdown>
-			</div>
-		);
-	}
-}
+	return (
+		<div>
+			<Heading>Click in the blank area of the viewport to dismiss the Dropdown</Heading>
+			<Dropdown
+				onClose={handleClose}
+				onOpen={handleOpen}
+				open={open} // initial value is true
+			>
+				{['test1', 'test2', 'test3']}
+			</Dropdown>
+		</div>
+	);
+};
 
-class DisabledDropdown extends Component {
-	constructor (props) {
-		super(props);
-		this.state = {
-			isDisabled: true
-		};
-	}
+const DisabledDropdown = () => {
+	const [isDisabled, setIsDisabled] = useState(true);
 
-	handleClick = () => {
-		if (this.state.isDisabled) {
-			this.setState({isDisabled: false});
-		} else {
-			this.setState({isDisabled: true});
-		}
-	};
+	const handleClick = useCallback(() => {
+		setIsDisabled(!isDisabled);
+	}, [isDisabled]);
 
-	render () {
-		return (
-			<div>
-				<Button onClick={this.handleClick}>{this.state.isDisabled ? 'Enable dropdown' : 'Disable dropdown'}</Button>
-				<Dropdown title="hello" disabled={this.state.isDisabled} onFocus={this.handleFocus}>
-					{['a', 'b', 'c']}
-				</Dropdown>
-			</div>
-		);
-	}
-}
+	return (
+		<div>
+			<Button onClick={handleClick}>{isDisabled ? 'Enable dropdown' : 'Disable dropdown'}</Button>
+			<Dropdown disabled={isDisabled} title="hello">
+				{['a', 'b', 'c']}
+			</Dropdown>
+		</div>
+	);
+};
 
 export default {
 	title: 'Agate/Dropdown',
 	component: 'Dropdown'
 };
 
-export const With2OptionsForTestingDirection = () => (
+export const With2OptionsForTestingDirection = (args) => (
 	<Dropdown
-		direction={select('direction', ['above', 'below'], Config)}
-		disabled={boolean('disabled', Config)}
+		direction={args['direction']}
+		disabled={args['disabled']}
 		onClose={action('onClose')}
 		onOpen={action('onOpen')}
 		onSelect={action('onSelect')}
 		style={{position: 'absolute', top: 'calc(50% - 4rem)'}}
-		title={text('title', Config, 'Dropdown')}
-		width={select('width', ['tiny', 'small', 'medium', 'large', 'x-large', 'huge'], Config)}
+		title={args['title']}
+		width={args['width']}
 	>
 		{['Option 1', 'Option 2']}
 	</Dropdown>
 );
 
+select('direction', With2OptionsForTestingDirection, ['above', 'below'], Config);
+boolean('disabled', With2OptionsForTestingDirection, Config);
+text('title', With2OptionsForTestingDirection, Config, 'Dropdown');
+select('width', With2OptionsForTestingDirection, ['smallest', 'small', 'medium', 'large', 'x-large', 'huge'], Config);
+
 With2OptionsForTestingDirection.storyName = 'with 2 options for testing direction';
 
-export const WithDefaultSelectedIn30Options = () => (
+export const WithDefaultSelectedIn30Options = (args) => (
 	<Dropdown
 		defaultSelected={10}
-		direction={select('direction', ['above', 'below'], Config)}
-		disabled={boolean('disabled', Config)}
+		direction={args['direction']}
+		disabled={args['disabled']}
 		onClose={action('onClose')}
 		onOpen={action('onOpen')}
 		onSelect={action('onSelect')}
-		title={text('title', Config, 'Dropdown')}
-		width={select('width', ['tiny', 'small', 'medium', 'large', 'x-large', 'huge'], Config)}
+		title={args['title']}
+		width={args['width']}
 	>
 		{items(30)}
 	</Dropdown>
 );
 
+select('direction', WithDefaultSelectedIn30Options, ['above', 'below'], Config);
+boolean('disabled', WithDefaultSelectedIn30Options, Config);
+text('title', WithDefaultSelectedIn30Options, Config, 'Dropdown');
+select('width', WithDefaultSelectedIn30Options, ['smallest', 'small', 'medium', 'large', 'x-large', 'huge'], Config);
+
 WithDefaultSelectedIn30Options.storyName = 'with defaultSelected in 30 options';
 
-export const WithLongText = () => (
+export const WithLongText = (args) => (
 	<Dropdown
-		direction={select('direction', ['above', 'below'], Config)}
-		disabled={boolean('disabled', Config)}
+		direction={args['direction']}
+		disabled={args['disabled']}
 		onClose={action('onClose')}
 		onOpen={action('onOpen')}
 		onSelect={action('onSelect')}
-		title={text('title', Config, 'Dropdown')}
-		width={select('width', ['tiny', 'small', 'medium', 'large', 'x-large', 'huge'], Config)}
+		title={args['title']}
+		width={args['width']}
 	>
 		{items(10, 'Looooooooooooooooooooooong')}
 	</Dropdown>
 );
 
+select('direction', WithLongText, ['above', 'below'], Config);
+boolean('disabled', WithLongText, Config);
+text('title', WithLongText, Config, 'Dropdown');
+select('width', WithLongText, ['smallest', 'small', 'medium', 'large', 'x-large', 'huge'], Config);
+
 WithLongText.storyName = 'with long text';
 
-export const WithMultipleDropdowns = () => (
+export const WithMultipleDropdowns = (args) => (
 	<div>
 		<Dropdown
-			direction={select('direction', ['above', 'below'], Config)}
-			disabled={boolean('disabled', Config)}
+			direction={args['direction']}
+			disabled={args['disabled']}
 			onClose={action('onClose')}
 			onOpen={action('onOpen')}
 			onSelect={action('onSelect')}
-			title={text('title', Config, 'Dropdown')}
-			width={select('width', ['tiny', 'small', 'medium', 'large', 'x-large', 'huge'], Config)}
+			title={args['title']}
+			width={args['width']}
 		>
 			{items(5)}
 		</Dropdown>
 		<Dropdown
-			direction={select('direction', ['above', 'below'], Config)}
-			disabled={boolean('disabled', Config)}
+			direction={args['direction']}
+			disabled={args['disabled']}
 			onClose={action('onClose')}
 			onOpen={action('onOpen')}
 			onSelect={action('onSelect')}
-			title={text('title', Config, 'Dropdown')}
-			width={select('width', ['tiny', 'small', 'medium', 'large', 'x-large', 'huge'], Config)}
+			title={args['title']}
+			width={args['width']}
 		>
 			{items(5)}
 		</Dropdown>
 	</div>
 );
 
+select('direction', WithMultipleDropdowns, ['above', 'below'], Config);
+boolean('disabled', WithMultipleDropdowns, Config);
+text('title', WithMultipleDropdowns, Config, 'Dropdown');
+select('width', WithMultipleDropdowns, ['smallest', 'small', 'medium', 'large', 'x-large', 'huge'], Config);
+
 WithMultipleDropdowns.storyName = 'with multiple dropdowns';
 
-export const WithArrayOfChildrenObjects = () => (
+export const WithArrayOfChildrenObjects = (args) => (
 	<div>
 		<Dropdown
-			direction={select('direction', ['above', 'below'], Config)}
-			disabled={boolean('disabled', Config)}
+			direction={args['direction']}
+			disabled={args['disabled']}
 			onClose={action('onClose')}
 			onOpen={action('onOpen')}
 			onSelect={action('onSelect')}
 			style={{position: 'absolute', top: 'calc(50% - 4rem)'}}
-			title={text('title', Config, 'Dropdown')}
-			width={select('width', ['tiny', 'small', 'medium', 'large', 'x-large', 'huge'], Config)}
+			title={args['title']}
+			width={args['width']}
 		>
 			{list}
 		</Dropdown>
 	</div>
 );
+
+select('direction', WithArrayOfChildrenObjects, ['above', 'below'], Config);
+boolean('disabled', WithArrayOfChildrenObjects, Config);
+text('title', WithArrayOfChildrenObjects, Config, 'Dropdown');
+select('width', WithArrayOfChildrenObjects, ['smallest', 'small', 'medium', 'large', 'x-large', 'huge'], Config);
 
 WithArrayOfChildrenObjects.storyName = 'with array of children objects';
 
@@ -189,13 +196,13 @@ export const WithDisabled = () => <DisabledDropdown />;
 
 WithDisabled.storyName = 'with disabled';
 
-export const WithGroupInScroller = () => {
-	const itemCount = number('items', Config, {range: true, min: 0, max: 50}, 5);
+export const WithGroupInScroller = (args) => {
+	const itemCount = args['items'];
 	const itemList = (new Array(itemCount)).fill().map((i, index) => `Option ${index + 1}`);
 	const dropdowns = [];
 
 	for (let i = 0; i < 30; i++) {
-		dropdowns.push({children: itemList, title: text('title', Config, 'Please select'), key: i});
+		dropdowns.push({children: itemList, title: args['title'], key: i});
 	}
 
 	return (
@@ -210,5 +217,8 @@ export const WithGroupInScroller = () => {
 		</Scroller>
 	);
 };
+
+range('items', WithGroupInScroller, Config, {range: true, min: 0, max: 50}, 5);
+text('title', WithGroupInScroller, Config, 'Please select');
 
 WithGroupInScroller.storyName = 'with group in Scroller';