+++ tests/screenshot/apps/components/DatePicker.js
@@ -1,16 +1,16 @@
 import DatePicker from '../../../../DatePicker';
 
 const DatePickerTests = [
-	<DatePicker value={new Date(2021, 6, 15)} />,
-	<DatePicker disabled value={new Date(2021, 6, 15)} />,
+	<DatePicker value={new Date(2022, 6, 30)} />,
+	<DatePicker disabled value={new Date(2022, 6, 30)} />,
 	// RTL
 	{
 		locale: 'ar-SA',
-		component: <DatePicker value={new Date(2021, 6, 15)} />
+		component: <DatePicker value={new Date(2022, 6, 30)} />
 	},
 	{
 		locale: 'ar-SA',
-		component: <DatePicker disabled value={new Date(2021, 6, 15)} />
+		component: <DatePicker disabled value={new Date(2022, 6, 30)} />
 	}
 ];