+++ tests/ui/specs/ArcPicker/ArcPicker-specs.js
@@ -5,54 +5,53 @@ const unselectedColor = 'rgb(136,136,136)';
 
 describe('ArcPicker', function () {
 
-	beforeEach(function () {
-		Page.open();
+	beforeEach(async function () {
+		await Page.open();
 	});
 
 	describe('default', function () {
 		const arcPicker = Page.components.arcPickerDefault;
 
-		it('should have the first arc selected by default', function () {
-			Page.spotlightSelect();
+		it('should have the first arc selected by default', async function () {
+			await Page.spotlightSelect();
 
-			expect(arcPicker.coloredPath(1).getCSSProperty('stroke').value).to.equal(accentColor);
+			expect((await arcPicker.coloredPath(1).getCSSProperty('stroke')).value).to.equal(accentColor);
 		});
 
-		it('should select the third arc when it is clicked', function () {
-			arcPicker.clickablePath(3).click();
-			expect(arcPicker.coloredPath(3).getCSSProperty('stroke').value).to.equal(accentColor);
+		it('should select the third arc when it is clicked', async function () {
+			await arcPicker.clickablePath(3).click();
+			expect((await arcPicker.coloredPath(3).getCSSProperty('stroke')).value).to.equal(accentColor);
 			// previous arcs should remain unselected
-			expect(arcPicker.coloredPath(1).getCSSProperty('stroke').value).to.equal(unselectedColor);
-			expect(arcPicker.coloredPath(2).getCSSProperty('stroke').value).to.equal(unselectedColor);
+			expect((await arcPicker.coloredPath(1).getCSSProperty('stroke')).value).to.equal(unselectedColor);
+			expect((await arcPicker.coloredPath(2).getCSSProperty('stroke')).value).to.equal(unselectedColor);
 		});
 	});
 
 	describe('cumulative', function () {
 		const arcPicker = Page.components.arcPickerCumulative;
 
-		it('should select all arcs up to the clicked arc', function () {
-			arcPicker.clickablePath(3).click();
-			expect(arcPicker.coloredPath(1).getCSSProperty('stroke').value).to.equal(accentColor);
-			expect(arcPicker.coloredPath(2).getCSSProperty('stroke').value).to.equal(accentColor);
-			expect(arcPicker.coloredPath(3).getCSSProperty('stroke').value).to.equal(accentColor);
+		it('should select all arcs up to the clicked arc', async function () {
+			await arcPicker.clickablePath(3).click();
+			expect((await arcPicker.coloredPath(1).getCSSProperty('stroke')).value).to.equal(accentColor);
+			expect((await arcPicker.coloredPath(2).getCSSProperty('stroke')).value).to.equal(accentColor);
+			expect((await arcPicker.coloredPath(3).getCSSProperty('stroke')).value).to.equal(accentColor);
 			// next one is unselected
-			expect(arcPicker.coloredPath(4).getCSSProperty('stroke').value).to.equal(unselectedColor);
+			expect((await arcPicker.coloredPath(4).getCSSProperty('stroke')).value).to.equal(unselectedColor);
 		});
 	});
 
 	describe('disabled', function () {
 		const arcPicker = Page.components.arcPickerDisabled;
 
-		it('should have the first arc selected by default', function () {
-			expect(arcPicker.coloredPath(1).getCSSProperty('stroke').value).to.equal(selectedColor);
+		it('should have the first arc selected by default', async function () {
+			expect((await arcPicker.coloredPath(1).getCSSProperty('stroke')).value).to.equal(selectedColor);
 		});
 
-		it('should not select the third arc when it is clicked', function () {
-			arcPicker.clickablePath(3).click();
-			expect(arcPicker.coloredPath(3).getCSSProperty('stroke').value).to.equal(unselectedColor);
+		it('should not select the third arc when it is clicked', async function () {
+			await arcPicker.clickablePath(3).click();
+			expect((await arcPicker.coloredPath(3).getCSSProperty('stroke')).value).to.equal(unselectedColor);
 			// first arc should remain selected
-			expect(arcPicker.coloredPath(1).getCSSProperty('stroke').value).to.equal(accentColor);
+			expect((await arcPicker.coloredPath(1).getCSSProperty('stroke')).value).to.equal(accentColor);
 		});
 	});
-
 });