+++ tests/ui/specs/Button/ButtonPage.js
@@ -11,12 +11,12 @@ class ButtonInterface {
 		this.badgeSelector = `#${this.id} > .enact_ui_Button_Button_decoration > .Button_Button_badge`;
 	}
 
-	focus () {
-		return browser.execute((el) => el.focus(), $(this.selector));
+	async focus () {
+		return await browser.execute((el) => el.focus(), await $(this.selector));
 	}
 
-	hover () {
-		return $(this.selector).moveTo({xOffset: 0, yOffset: 0});
+	async hover () {
+		return await $(this.selector).moveTo({xOffset: 0, yOffset: 0});
 	}
 
 	get self () {
@@ -90,8 +90,8 @@ class ButtonPage extends Page {
 		};
 	}
 
-	open (urlExtra) {
-		super.open('Button-View', urlExtra);
+	async open (urlExtra) {
+		await super.open('Button-View', urlExtra);
 	}
 }