+++ tests/ui/specs/Drawer/Drawer-utils.js
@@ -1,22 +1,22 @@
 // Utility methods for testing
 
-function validateTitle (drawer, title) {
-	expect(drawer.title).to.equal(title);
+async function validateTitle (drawer, title) {
+	expect(await drawer.title).to.equal(title);
 }
 
-function expectClosed (drawer) {
-	expect(drawer.isDrawerExist).to.be.false();
-	expect(drawer.isScrimExist).to.be.false();
+async function expectClosed (drawer) {
+	expect(await drawer.isDrawerExist).to.be.false();
+	expect(await drawer.isScrimExist).to.be.false();
 }
 
-function expectOpen (drawer) {
-	expect(drawer.isDrawerExist).to.be.true();
-	expect(drawer.isScrimExist).to.be.true();
+async function expectOpen (drawer) {
+	expect(await drawer.isDrawerExist).to.be.true();
+	expect(await drawer.isScrimExist).to.be.true();
 }
 
-function expectNoneScrimOpen (drawer) {
-	expect(drawer.isDrawerExist).to.be.true();
-	expect(drawer.isScrimExist).to.be.false();
+async function expectNoneScrimOpen (drawer) {
+	expect(await drawer.isDrawerExist).to.be.true();
+	expect(await drawer.isScrimExist).to.be.false();
 }
 
 module.exports = {