+++ tests/ui/specs/Popup/PopupPage.js
@@ -93,28 +93,28 @@ class PopupPage extends Page {
 		this.components.popup10 = new PopupInterface('popup10');
 	}
 
-	open (urlExtra) {
-		super.open('Popup-View', urlExtra);
+	async open (urlExtra) {
+		await super.open('Popup-View', urlExtra);
 	}
 
-	clickPopupFloatLayer () {
-		$('#floatLayer').click();
+	async clickPopupFloatLayer () {
+		await $('#floatLayer > div').click({y: -400});
 	}
 
-	waitForOpen (selector, timeout) {
+	async waitForOpen (selector, timeout) {
 		if (typeof selector !== 'string') selector = `#${selector.id}`;
 
-		$(selector).waitForExist({timeout});
+		await $(selector).waitForExist({timeout});
 	}
 
-	waitForClose (selector, timeout) {
+	async waitForClose (selector, timeout) {
 		if (typeof selector !== 'string') selector = `#${selector.id}`;
 
-		$(selector).waitForExist({timeout, reverse: true});
+		await $(selector).waitForExist({timeout, reverse: true});
 	}
 
-	waitForFocused (node, timeout, timeoutMsg = 'timed out waiting for focus', interval = 250) {
-		browser.waitUntil(function () {
+	async waitForFocused (node, timeout, timeoutMsg = 'timed out waiting for focus', interval = 250) {
+		await browser.waitUntil(function () {
 			return node.isFocused();
 		}, {timeout, timeoutMsg, interval});
 	}