+++ tests/ui/specs/SliderButton/SliderButton-specs.js
@@ -1,37 +1,37 @@
 const Page = require('./SliderButtonPage');
 
 describe('SliderButton', function () {
-	beforeEach(function () {
-		Page.open();
+	beforeEach(async function () {
+		await Page.open();
 	});
 
 	describe('slider button with three items', function () {
 		const sliderButton = Page.components.sliderButtonThreeItems;
 
-		it('should have focus on first button at start', function () {
-			expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0');
+		it('should have focus on first button at start', async function () {
+			expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0');
 		});
 
-		it('should have focus on second button when clicked on `Ridiculous Speed`', function () {
-			sliderButton.clickableItem(1).click();
+		it('should have focus on second button when clicked on `Ridiculous Speed`', async function () {
+			await sliderButton.clickableItem(1).click();
 
-			expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0.5');
+			expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0.5');
 		});
 
-		it('should have focus on third button when clicked `Ludicrous Speed`', function () {
-			sliderButton.clickableItem(2).click();
+		it('should have focus on third button when clicked `Ludicrous Speed`', async function () {
+			await sliderButton.clickableItem(2).click();
 
-			expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('1');
+			expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('1');
 		});
 
 		describe('5-way', function () {
-			it('should change value between options on 5-way navigation', function () {
-				Page.spotlightRight();
-				expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0.5');
-				Page.spotlightRight();
-				expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('1');
-				Page.spotlightLeft();
-				expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0.5');
+			it('should change value between options on 5-way navigation', async function () {
+				await Page.spotlightRight();
+				expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0.5');
+				await Page.spotlightRight();
+				expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('1');
+				await Page.spotlightLeft();
+				expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0.5');
 			});
 		});
 	});
@@ -39,50 +39,50 @@ describe('SliderButton', function () {
 	describe('slider button with five items', function () {
 		const sliderButton = Page.components.sliderButtonFiveItems;
 
-		it('should have focus on first button at start', function () {
-			expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0');
+		it('should have focus on first button at start', async function () {
+			expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0');
 		});
 
-		it('should have focus on second item when clicked on `Ridiculous Speed`', function () {
-			sliderButton.clickableItem(1).click();
+		it('should have focus on second item when clicked on `Ridiculous Speed`', async function () {
+			await sliderButton.clickableItem(1).click();
 
-			expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0.25');
+			expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0.25');
 		});
 
-		it('should have focus on third item when clicked on `Ludicrous Speed`', function () {
-			sliderButton.clickableItem(2).click();
+		it('should have focus on third item when clicked on `Ludicrous Speed`', async function () {
+			await sliderButton.clickableItem(2).click();
 
-			expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0.5');
+			expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0.5');
 		});
 
-		it('should have focus on forth item when clicked on `Bananas Speed`', function () {
-			sliderButton.clickableItem(3).click();
+		it('should have focus on forth item when clicked on `Bananas Speed`', async function () {
+			await sliderButton.clickableItem(3).click();
 
-			expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0.75');
+			expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0.75');
 		});
 
-		it('should have focus on fifth item when clicked on `OK Enough Speed`', function () {
-			sliderButton.clickableItem(4).click();
+		it('should have focus on fifth item when clicked on `OK Enough Speed`', async function () {
+			await sliderButton.clickableItem(4).click();
 
-			expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('1');
+			expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('1');
 		});
 
 		describe('5-way', function () {
-			it('should change value between options on 5-way navigation', function () {
+			it('should change value between options on 5-way navigation', async function () {
 				// first move focus on current sliderButton
-				Page.spotlightDown();
+				await Page.spotlightDown();
 				// try to move focus to second option
-				Page.spotlightRight();
-				expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0.25');
-				Page.spotlightRight();
-				expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0.5');
-				Page.spotlightRight();
-				expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0.75');
-				Page.spotlightRight();
-				expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('1');
-				Page.spotlightLeft();
-				Page.spotlightLeft();
-				expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0.5');
+				await Page.spotlightRight();
+				expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0.25');
+				await Page.spotlightRight();
+				expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0.5');
+				await Page.spotlightRight();
+				expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0.75');
+				await Page.spotlightRight();
+				expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('1');
+				await Page.spotlightLeft();
+				await Page.spotlightLeft();
+				expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0.5');
 			});
 		});
 	});
@@ -90,24 +90,24 @@ describe('SliderButton', function () {
 	describe('disabled', function () {
 		const sliderButton = Page.components.sliderButtonDisabled;
 
-		it('should have focus on first item at start', function () {
-			expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0');
+		it('should have focus on first item at start', async function () {
+			expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0');
 		});
 
-		it('should not move focus when clicked on second item', function () {
-			sliderButton.clickableItem(2).click();
+		it('should not move focus when clicked on second item', async function () {
+			await sliderButton.clickableItem(2).click();
 
-			expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0');
+			expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0');
 		});
 
 		describe('5-way', function () {
-			it('should not change value between options on 5-way navigation', function () {
+			it('should not change value between options on 5-way navigation', async function () {
 				// first move focus on current sliderButton
-				Page.spotlightDown();
-				Page.spotlightDown();
+				await Page.spotlightDown();
+				await Page.spotlightDown();
 				// try to move focus to second option
-				Page.spotlightRight();
-				expect(sliderButton.getKnob().getAttribute('proportion')).to.equal('0');
+				await Page.spotlightRight();
+				expect(await sliderButton.getKnob().getAttribute('proportion')).to.equal('0');
 			});
 		});
 	});