+++ CHANGELOG.md
@@ -6,19 +6,48 @@ The following is a curated list of changes in the Enact agate module, newest cha
 
 ### Added
 
-- `agate/Heading` prop `showBackButton`
-- `agate/Input` prop `clearButton` and `clearIcon`
-- `agate/Keypad` prop `activeCall` to render different icons depending on whether there is an active call or not
+- `agate/Header` Marquee for title, subtitle, and titleAbove
+
+### Deprecated
+
+- `agate/GridListImageItem`, to be removed in 2.0.0. Use `agate/ImageItem` instead
 
 ### Fixed
 
-- `agate/Heading` style to match latest design for Silicon skin
-- `agate/Input` to match latest design for Silicon skin
-- `agate/LabeledIconButton` to match the latest design for Silicon skin
-- `agate/Keypad` to match latest design for Silicon skin
+- `agate/ContextualPopupDecorator` layout for Carbon, Cobalt, Copper, Electro, Titanium skins
+- `agate/IncrementSlider` style to have a proper layout for every skin
+- `agate/MediaPlayer` layout issues for Cobalt, Carbon, Copper, Electro, Titanium skins
+- `agate/MediaPlayer` shuffle button color for Cobalt, and Copper skins
+- `agate/Panel` and `agate/TabGroup` padding for RTL locales
 - `agate/Picker` and `agate/RangePicker` to match latest design for Silicon skin
+- `agate/PopupMenu` title to marquee for long text
+- `agate/TabGroup` button padding for Cobalt and Copper skins
+- `agate/VirtualList` 5-way navigation between scroll buttons when `focusableScrollbar`
 
-## [2.0.0-beta.3] - 2021-06-06
+## [2.0.0-beta.2] - 2021-06-24
+
+### Added
+
+- `agate/Heading` prop `showBackButton`
+- `agate/Input` prop `clearButton` and `clearIcon`
+- `agate/Keypad` prop `activeCall` to render different icons depending on whether there is an active call or not
+- `agate/MediaPlayer` prop `type`
+
+### Fixed
+
+- `agate/Dropdown` style to match latest design for Silicon skin
+- `agate/Heading` style to match latest design for Silicon skin
+- `agate/Heading` position of underline when size="tile" on Carbon, Cobalt, Copper, Electro, Titanium skins
+- `agate/Input` to match latest design for Silicon skin
+- `agate/Item` for inline with label and labelPosition "after" or "before" on Gallium skin to not change item width on hover
+- `agate/Keypad` to match latest design for Silicon skin
+- `agate/LabeledIconButton` to match the latest design for Silicon skin
+- `agate/MediaPlayer` style to match latest design for Silicon skin
+- `agate/Slider` height when `orientation="vertical"` on Carbon, Cobalt, Copper, Electro, Titanium skins
+- `agate/SwitchItem` style to match latest design for Silicon skin
+- `agate/TooltipDecorator` to to match latest design for Silicon skin
+
+## [2.0.0-beta.1] - 2021-06-06
 
 ### Added